// frontend/src/lib/api/model/questionCard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * e-proshno API
 * OpenAPI spec version: v1
 */
import type { ContentStatus } from './contentStatus';
import type { CqPartDto } from './cqPartDto';
import type { McqKind } from './mcqKind';
import type { OptionDto } from './optionDto';
import type { QuestionType } from './questionType';

export interface QuestionCard {
  id: string;
  type: QuestionType;
  mcqKind?: McqKind | null;
  stem: string;
  /** @nullable */
  stimulusId?: string | null;
  /** @nullable */
  stimulus?: string | null;
  options: OptionDto[];
  cqParts: CqPartDto[];
  /** @nullable */
  explanation?: string | null;
  boardTags: string[];
  difficulty: number;
  importance: number;
  isMath: boolean;
  hasImage: boolean;
  isCommon: boolean;
  /** @nullable */
  bankId?: string | null;
  /** @nullable */
  bankName?: string | null;
  subjectId: string;
  chapterId: string;
  chapterNumber: number;
  chapterName: string;
  /** @nullable */
  topicName?: string | null;
  status: ContentStatus;
  canEdit: boolean;
  updatedAt: string;
}